dev: Setup database, migrations and test setup
Prep bastion to use a Postgres database for storing information.
dropwizard-jdbi3 is used as the ORM.
dropwizard-flyway is used for schema migrations. The module had to
be physically copied as a new version has not been released.

A simple test that brings up the application and setups the database is
also available. Developer workflow has to be documented. circleci is
setup to provide a database for testing.  

Yarn and Node has also been cached to improve build time.