feature: Add an API to register an organization and root user
A new api to bootstrap an installation has been added that will accept
the basic information for an org and user. The user is considered to
have root privileges and can be used for further setup.

Major changes were also made to handling passwords in the application
and database. passwords are stored in bytea in database and byte[] in
the application.