new: Support refresh token to refresh JWT tokens
Reference: https://blog.hasura.io/best-practices-of-using-jwt-with-graphql/

Backend creates a secure random string as a refresh token when user logs
in. The token is stored in the database. It is sent to the client in a
cookie along with the rest of the response after successful login.

A new auth filter has been added. It used to authenticate to
/refreshJWT.